requirements.txt
tensorflow>=2.10.0
pandas>=1.5.3
numpy>=1.23.5
boto3>=1.26.137
scikit-learn>=1.2.2
python-dotenv>=1.0.0
matplotlib>=3.7.1
seaborn>=0.12.2
